Bug756:Merged SAS/OTB release branch into trunk again

+++ trunk/SAS/OTB/OTB/src/nl/astron/lofar/sas/otb/panels/TemplateMaintenancePanel.java
@@ -34,7 +34,6 @@ import javax.swing.tree.TreePath;
 import nl.astron.lofar.lofarutils.LofarUtils;
 import nl.astron.lofar.sas.otb.MainFrame;
 import nl.astron.lofar.sas.otb.jotdb3.jOTDBnode;
-import nl.astron.lofar.sas.otb.jotdb3.jOTDBparam;
 import nl.astron.lofar.sas.otb.util.ConfigPanelHelper;
 import nl.astron.lofar.sas.otb.util.IViewPanel;
 import nl.astron.lofar.sas.otb.util.OtdbRmi;
@@ -226,12 +225,33 @@ public class TemplateMaintenancePanel extends javax.swing.JPanel
         logger.debug("actionPerformed: " + evt);
         logger.debug("Trigger: " + evt.getActionCommand());
         if (evt.getActionCommand().equals("Delete")) {
+
             //Check  if the selected node isn't a leaf
-            if (itsSelectedNode != null && !itsSelectedNode.leaf) {
+            if (itsSelectedNode != null && !itsSelectedNode.leaf && itsSelectedNode.instances <= 1 ) {
                 if (JOptionPane.showConfirmDialog(this, "Are you sure you want to delete this node ?", "Delete Tree", JOptionPane.YES_NO_OPTION, JOptionPane.WARNING_MESSAGE) == JOptionPane.YES_OPTION) {
                     try {
                         if (OtdbRmi.getRemoteMaintenance().deleteNode(itsSelectedNode)) {
                             logger.debug("Node + children deleted");
+
+                          
+                            // We have to find the defaultNode for this deleted node and decrease the number of instances
+                            Vector<jOTDBnode> aList = OtdbRmi.getRemoteMaintenance().getItemList(itsTreeID, itsSelectedNode.name);
+                            Iterator<jOTDBnode> it = aList.iterator ();
+                            jOTDBnode aDefaultNode=null;
+                            short maxIdx=0;
+                            while (it.hasNext ()) {
+                                jOTDBnode aN = it.next ();
+                                if (aN.index > maxIdx) maxIdx=aN.index;
+                                if (aN.index == -1) {
+                                   aDefaultNode=aN;
+                                }
+                            }
+                            if (aDefaultNode != null) {
+                                maxIdx+=1;
+                                // Grootste index of totaal aantal
+                                aDefaultNode.instances=maxIdx;
+                                OtdbRmi.getRemoteMaintenance().saveNode(aDefaultNode);
+                            }
                             setNewRootNode();
                         }
                     } catch (RemoteException ex) {
@@ -241,12 +261,8 @@ public class TemplateMaintenancePanel extends javax.swing.JPanel
             }
         } else if (evt.getActionCommand().equals("Duplicate")) {
 
-            //!!!!!  if duplicating a node the default node.instances need to be set to +1 instance!!!!!!
-            itsMainFrame.ToDo();
-
-
-            //Check  if the selected node isn't a leaf
-            if (itsSelectedNode != null && !itsSelectedNode.leaf) {
+            //Check  if the selected node isn't a leaf and it is a default node (index = -1)
+            if (itsSelectedNode != null && !itsSelectedNode.leaf && itsSelectedNode.index==-1) {
                 String answer = JOptionPane.showInputDialog(this, "What is the index for the new subtree?", "Enter indexNumber", JOptionPane.QUESTION_MESSAGE);
                 if (answer != null || !answer.equals("")) {
                     short idx = Integer.valueOf(answer).shortValue();
@@ -258,7 +274,13 @@ public class TemplateMaintenancePanel extends javax.swing.JPanel
                         int aN = OtdbRmi.getRemoteMaintenance().dupNode(itsTreeID, itsSelectedNode.nodeID(), idx);
                         if (aN > 0) {
                             logger.debug("Node duplicated");
-                            setNewRootNode();
+                            // defaultNode.instances needs 2b set to highest instance
+                           if (idx+1 > itsSelectedNode.instances) {
+                               Integer newInst=idx+1;
+                               itsSelectedNode.instances=newInst.shortValue();
+                               OtdbRmi.getRemoteMaintenance().saveNode(itsSelectedNode);
+                               setNewRootNode();
+                           }
                         } else {
                             logger.debug("Node duplication failed");
                         }
@@ -389,7 +411,7 @@ public class TemplateMaintenancePanel extends javax.swing.JPanel
         int savedSelection = jTabbedPane1.getSelectedIndex();
         logger.debug("ChangeSelection for node: " + aNode.name);
         itsSelectedNode = aNode;
-        jOTDBparam aParam = null;
+
 
         changeSelection(aNode);
 
@@ -399,12 +421,33 @@ public class TemplateMaintenancePanel extends javax.swing.JPanel
 
         } else {
             int[] selectedRows = treePanel.getSelectedRows();
-            if (selectedRows == null || selectedRows.length <= 0 || selectedRows[0] == 0 || aNode.leaf) {
+            if (selectedRows == null || selectedRows.length <= 0 || selectedRows[0] == 0 || aNode.leaf ) {
                 buttonPanel1.setButtonEnabled("Duplicate", false);
                 buttonPanel1.setButtonEnabled("Delete", false);
             } else {
-//                buttonPanel1.setButtonEnabled("Duplicate",true);
-                buttonPanel1.setButtonEnabled("Delete", true);
+                // only duplication possible on DefaultTrees (index = -1)
+                if (aNode.index == -1) {
+                    buttonPanel1.setButtonEnabled("Duplicate",true);
+                } else {
+                    buttonPanel1.setButtonEnabled("Duplicate",false);
+                }
+                try {
+                    // only deletion possible when no instances left
+                    Vector<jOTDBnode> aList = OtdbRmi.getRemoteMaintenance().getItemList(itsTreeID, itsSelectedNode.name);
+
+                    if (aNode.index == -1 && aList != null && aList.size() <= 1 ) {
+                        buttonPanel1.setButtonEnabled("Delete", true);
+                    } else if (aNode.index != -1 && aNode.instances == 1) {
+                        buttonPanel1.setButtonEnabled("Delete", true);
+                    } else {
+                        buttonPanel1.setButtonEnabled("Delete", false);
+                    }
+               } catch (RemoteException ex) {
+                    logger.error("Error: Couldn't get ItemList" , ex);
+                    buttonPanel1.setButtonEnabled("Delete", true);
+                    return;
+                }
+
             }
         }
         if (savedSelection > -1 && savedSelection < jTabbedPane1.getComponentCount()) {
@@ -426,7 +469,6 @@ public class TemplateMaintenancePanel extends javax.swing.JPanel
             buttonPanel1.addButton("Delete");
             buttonPanel1.addButton("Duplicate");
             buttonPanel1.addButton("Exit");
-            buttonPanel1.setButtonEnabled("Duplicate", false);
         }
